Create README.md
This is a Fortran simulation of a one-dimensional plasma sheath as described in "Plasma Physics via Computer Simulation" by Birdsall and Langdon. The simulation consists of a plasma bounded by two surfaces with a particle source in the center of the domain. When an ion arrives at a boundary an electron-ion pair is placed at the source, thus ensuring a rapid steady-state.